+++ dolby_vision/src/rpu/dovi_rpu.rs
@@ -51,7 +51,7 @@ pub struct DoviRpu {
             skip_serializing_if = "BitVec::is_empty"
         )
     )]
-    pub remaining: BitVec<Msb0, u8>,
+    pub remaining: BitVec<u8, Msb0>,
     pub rpu_data_crc32: u32,
 
     #[cfg_attr(feature = "serde_feature", serde(skip_serializing))]